+++ src/secplugin/__init__.py
@@ -1,6 +1,6 @@
 import warnings
 
-from secplugin.plugin import Plugin
+from secplugin.plugin import ConcurrencyMode, Plugin
 from secplugin.messenger import Messenger
 from secplugin.cmd import Cmd
 from secplugin.msg import Msg
@@ -8,5 +8,5 @@ from secplugin.sender import Sender
 
 warnings.filterwarnings("default", category=DeprecationWarning, module=__name__)
 
-__version__ = "1.2.6"
-__all__ = ["Plugin", "Messenger", "Cmd", "Msg", "Sender"]
+__version__ = "1.3.1"
+__all__ = ["ConcurrencyMode", "Plugin", "Messenger", "Cmd", "Msg", "Sender"]