+++ src/secplugin/cmd.py
@@ -1,8 +1,8 @@
 from enum import Enum
 
 class Cmd(str, Enum):
-    SyncOicq = "SyncOicq"
-    Response = "Response"
-    Heartbeat = "Heartbeat"
+    Sync        = "Sync"
+    Response    = "Response"
+    Heartbeat   = "Heartbeat"
     PushOicqMsg = "PushOicqMsg"
     SendOicqMsg = "SendOicqMsg"